+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 ### 0.1.0 (25.06.2021)
     - initial release
+### 0.1.1 (27.08.2021)
+    - Fix the IDS colliding at litello's side(email already registered error)
+    - Add a Tile image(Kachelbild)
+### 0.1.2 (06.04.22)
+    - Fix Permissions: Users with Write Access are not directly redirected to content page when they open the book